[WIP]Admin api to provide cluster stats
Currently there is no admin API for Magnum to get the operator
view of the system. This change is to add an admin API
endpoint that can provide following basic information;
1) Total number of clusters and cluster nodes across all tenants.
2) Total clusters and nodes for a given tenant

This admin API will be accessible to users with 'admin' role.

NOTE: Specification for this work is in progress.

Change-Id: Iac0bf9343549de31654545d5b1fd7601e56142a7
Closes-Bug: #1634577